Add allocation queue creation to event log
