+++ crates/hyperqueue/src/server/event/events.rs
@@ -1,3 +1,5 @@
+use crate::server::autoalloc::DescriptorId;
+use crate::transfer::messages::AllocationQueueParams;
 use crate::WorkerId;
 use serde::{Deserialize, Serialize};
 use tako::messages::common::WorkerConfiguration;
@@ -7,9 +9,13 @@ use tako::static_assert_size;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub enum MonitoringEventPayload {
+    // Workers
     WorkerConnected(WorkerId, Box<WorkerConfiguration>),
     WorkerLost(WorkerId, LostWorkerReason),
-    OverviewUpdate(WorkerOverview),
+    WorkerOverviewReceived(WorkerOverview),
+    // Allocations
+    AllocationQueueCreated(DescriptorId, Box<AllocationQueueParams>),
+    AllocationQueueRemoved(DescriptorId),
 }
 
 // Keep the size of the event structure in check
